QCReport(affyQCReport)	R Documentation

QCReport

Description

Creates a QC report for an object of class AffyBatch.

Usage


QCReport(object,file="AffyQCReport.pdf",...)

Arguments

object 	 An object of class AffyBatch.

file 	 A valid filename for the output PDF file.

... 	 further arguments for plot function.

Details

This creates a 6 page QC report in PDF file format.

The default output file is AffyQCReport.pdf created in the working directory.
The AffyBatch object is not required to contain phenotypic data but that data is used if available.
The CDF environment must be available for the method to work fully.

More details are available in the vignette.

Value

TRUE

Examples


  ## Not run: 
     
    library(affydata)
    data(Dilution)
    
  

    QCReport(Dilution,file="ExampleQC.pdf")   ##A QC report for the Example data will be generated
  ## End(Not run)

[Package affyQCReport version 1.6-1 Index]