borderQC1(affyQCReport)	R Documentation

Distribution of intensities of the border elements

Description

Creates the forth page of the QC report created by QCReport for an object of class AffyBatch The page contains a boxplots showing the intensities from the border element of arrays in an AffyBatch object.

Usage


borderQC1(object)

Arguments

object

{An object of class AffyBatch .}

Details

This creates the forth page of a QCReport report. Boxplots of the positive and negative border elements of arrays in a AffyBatch object are shown.

More details are available in the vignette.

Value

TRUE

Examples


    
    library(affydata)
    data(Dilution)

   borderQC1(Dilution)


[Package affyQCReport version 1.6-2 Index]