correlationPlot(affyQCReport)	R Documentation

Array-array intensity correlation plot

Description

Creates the sixth page of the QC report created by QCReport for an object of class AffyBatch The page displays array-array intensity correlation coefficients.

Usage


correlationPlot(object)

Arguments

object

{An object of class AffyBatch.}

Details

This creates the sixth page of a QCReport report. The page displays array-array intensity correlation coefficients for all pairs of arrays in an AffyBatch object. Arrays are ordered according to phenotypic data if available. This plot is useful for detecting outlier arrays.

More details are available in the vignette.

Value

TRUE

Examples



   library(affydata)
    
   data(Dilution)

  correlationPlot(Dilution)


[Package affyQCReport version 1.6-2 Index]